requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
langsmith>=0.0.69
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.6.0
jupyter>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2